# src/infrastructure/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Simon Backend API",
    "description": "Documentación de la API de Simon Backend"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
...
    "/reserve/restaurant/{restauranteId}": {
      "get": {
        "summary": "Obtiene todas las reservas de un restaurante específico",
        "parameters": [
          {
            "name": "restauranteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas del restaurante"
          }
        }
      }
    },
...
    "/dish/restaurant/{restauranteId}": {
      "get": {
        "summary": "Obtiene todos los platillos de un restaurante específico",
        "parameters": [
          {
            "name": "restauranteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de platillos del restaurante"
          }
        }
      }
    },
...
    "/restaurant/{restaurantId}/img": {
      "get": {
        "summary": "Obtiene las imágenes de un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de imágenes"
          }
        }
      }
    },
    "/restaurant/{restaurantId}/images/{type}": {
      "patch": {
        "summary": "Actualiza múltiples imágenes de un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Tipo de imagen"
          }
        ],
        "responses": {
          "200": {
            "description": "Imágenes actualizadas"
          }
        }
      },
      "post": {
        "summary": "Sube múltiples imágenes a un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Tipo de imagen"
          }
        ],
        "responses": {
          "200": {
            "description": "Imágenes subidas"
          }
        }
      }
    },
    "/restaurant/{restaurantId}/img/{type}": {
      "post": {
        "summary": "Sube una imagen a un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Tipo de imagen"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen subida"
          }
        }
      },
      "put": {
        "summary": "Actualiza una imagen de un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          },
          {
            "name": "imgUrl",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "URL de la imagen"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen actualizada"
          }
        }
      },
      "delete": {
        "summary": "Elimina una imagen de un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          },
          {
            "name": "imgId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de la imagen"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen eliminada"
          }
        }
      }
    },
    "/restaurant/{restaurantId}": {
      "get": {
        "summary": "Obtiene un restaurante por ID",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del restaurante"
          }
        }
      },
      "patch": {
        "summary": "Actualiza un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurante actualizado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Restaurante eliminado"
          }
        }
      }
    },
...
    "/restaurant/{restaurantId}/suscription": {
      "patch": {
        "summary": "Actualiza la suscripción de un restaurante",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del restaurante"
          }
        ],
        "responses": {
          "200": {
            "description": "Suscripción actualizada"
          }
        }
      }
    },
...
  }
}